Pin everything to the new 22.05 release
Previously Bugg (the Mac) used the unstable releases of nixpkgs and
home-manager. Around the release of 22.05 I was getting warnings because
nixpkgs (unstable) had moved to track 22.05 while home-manager was still
reporting that it was 21.11.

Since 22.05 is up-to-date I'll now instead pin Bugg to that version and
expose a pkgs.unstable overlay to allow specific packages to be
installed from unstable.

Signed-off-by: Nic Cope <nicc@rk0n.org>